requirements.txt
chardet==3.0.4
idna==2.10
urllib3==1.26.3
appdirs==1.4.3
Cython==0.23
Django==1.10.6
docutils==0.13.1
openpyxl==3.0.6
packaging==16.8
pipenv==11.8.0
psutil==5.0.1
pyowm==2.8.0
Pygments==2.2.0
pyparsing==2.2.0
python-telegram-bot==13.1
six==1.15.0
virtualenv==20.2.0
